Fix following splint warning.
src/tree.c: (in function rewrite_p)
src/tree.c:2092:16: Operands of < have incompatible types (unsigned int,
                       size_t): i < a->count

+++ src/tree.c
@@ -2058,7 +2058,7 @@ static void rewrite_a(struct tree_array *a, struct rewrite_ctx *ctx)
 
 static void rewrite_p(struct param_array *a, struct rewrite_ctx *ctx)
 {
-   for (unsigned i = 0; i < a->count; i++) {
+   for (size_t i = 0; i < a->count; i++) {
       switch (a->items[i].kind) {
       case P_RANGE:
          a->items[i].range.left =
